fix: rm evil-collection from overrides
This change was added to allow evil-collection to compile on latest
versions macOS Ventura, the evil-collection version at the time of this
change was halting on the compilation of the
`evil-collection-speedbar.el` file.

Unfortunately, with this approach, although it finishes the nix-doom
emacs compilation, leaves doom-emacs' runtime in a faulty state. The
emacs package relies on a sub-directory `modes` that doesn't get
compiled when specifying this dependency in the overrides list.

As a way to make it work as expected on macOS Ventura, users should
override the `emacsPackagesOverlay` option and specify the
`evil-collection` variable. Following an example

```
{ pkgs, ... }:
{
  programs.doom-emacs = {
    enable = true;
    emacsPackagesOverlay = _final: _prev: {
        inherit (pkgs.emacsPackages) evil-collection;
    };
  };
}
```